LPS-73999 Package name should match the Bundle-SymbolicName specified in the bnd settings
