+++ modules/apps/forms-and-workflow/portal-rules-engine/portal-rules-engine-sample-web/docroot/WEB-INF/src/com/liferay/portal/rules/engine/sample/dependencies/rules_user_custom_attribute_content.drl
