+++ modules/apps/forms-and-workflow/portal-rules-engine/portal-rules-engine-sample-web/docroot/WEB-INF/src/com/liferay/portal/rules/engine/sample/web/ConfigurationActionImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sampledrools.action;
+package com.liferay.portal.rules.engine.sample.web;
 
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
