+++ modules/apps/web-experience/layout/layout-impl/src/main/java/com/liferay/layout/internal/model/adapter/StagedLayoutSetImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.set.internal.model.adapter;
+package com.liferay.layout.internal.model.adapter;
 
 import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.exportimport.kernel.lar.StagedModelType;
