+++ modules/apps/web-experience/layout/layout-impl/src/main/java/com/liferay/layout/internal/model/adapter/builder/LayoutSetModelAdapterBuilder.java
@@ -12,25 +12,38 @@
  * details.
  */
 
-package com.liferay.layout.set.internal.model.adapter.builder;
+package com.liferay.layout.internal.model.adapter.builder;
 
-import com.liferay.layout.set.internal.model.adapter.StagedLayoutSetImpl;
+import com.liferay.layout.internal.model.adapter.StagedLayoutSetImpl;
 import com.liferay.layout.set.model.adapter.StagedLayoutSet;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.model.adapter.builder.ModelAdapterBuilder;
+import com.liferay.portal.kernel.util.ReflectionUtil;
+
+import java.lang.reflect.Field;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Mate Thurzo
+ * @author Daniel Kocsis
  */
 @Component(immediate = true, service = ModelAdapterBuilder.class)
-public class StagedLayoutSetModelAdapterBuilder
-	implements ModelAdapterBuilder<LayoutSet, StagedLayoutSet> {
+public class LayoutSetModelAdapterBuilder
+	implements ModelAdapterBuilder<StagedLayoutSet, LayoutSet> {
 
 	@Override
-	public StagedLayoutSet build(LayoutSet layoutSet) {
-		return new StagedLayoutSetImpl(layoutSet);
+	public LayoutSet build(StagedLayoutSet stagedLayoutSet) {
+		Field layoutSetField = null;
+
+		try {
+			layoutSetField = ReflectionUtil.getDeclaredField(
+				StagedLayoutSetImpl.class, "_layoutSet");
+
+			return (LayoutSet)layoutSetField.get(stagedLayoutSet);
+		}
+		catch (Exception e) {
+			return null;
+		}
 	}
 
 }
