+++ modules/apps/web-experience/layout/layout-impl/src/main/java/com/liferay/layout/internal/xstream/configurator/LayoutXStreamConfigurator.java
@@ -17,7 +17,7 @@ package com.liferay.layout.internal.xstream.configurator;
 import com.liferay.exportimport.kernel.xstream.XStreamAlias;
 import com.liferay.exportimport.kernel.xstream.XStreamConverter;
 import com.liferay.exportimport.kernel.xstream.XStreamType;
-import com.liferay.layout.set.internal.model.adapter.StagedLayoutSetImpl;
+import com.liferay.layout.internal.model.adapter.StagedLayoutSetImpl;
 import com.liferay.layout.set.model.adapter.StagedLayoutSet;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.model.LayoutSetStagingHandler;
