+++ modules/core/portal-app-license-resolver-hook/src/main/java/com/liferay/portal/app/license/resolver/hook/AppResolverHook.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.app.license.resolver;
+package com.liferay.portal.app.license.resolver.hook;
 
 import com.liferay.portal.app.license.AppLicenseVerifier;
 import com.liferay.portal.kernel.log.Log;
